styling

+++ db/migrate/20121229161418_create_pesees.rb
@@ -2,7 +2,7 @@ class CreatePesees < ActiveRecord::Migration
   def change
     create_table :pesees do |t|
       t.date :quand
-      t.integer :what
+      t.float :what
       t.integer :user_id
 
       t.timestamps
