+++ db/migrate/20130102083703_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration
   def change
     create_table :ingredients do |t|
       t.integer :aliment_id
-      t.integer :quantite
+      t.float :quantite
       t.integer :repa_id
       t.timestamps
     end
